improve raycast snapping behavior
When snapping to a vertex, there is often a vertex of a different part at the same location.
The same goes for edges.
This commit adds a method for finding the most suitable of these vertices/edges, by checking
for which of the parent objects the ray intersects with a face of that object.